backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build-backend"

[project]
name = "nexus-backend"
version = "0.1.0"
description = "Scalable AI Agent Service for internal documentation querying"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Rupert Cobbe-Warburton", email = "rupertcw10@gmail.com"}]

dependencies = [
    "fastapi==0.111.*",
    "uvicorn==0.30.*",
    "sqlalchemy==2.0.*",
    "psycopg2==2.9.*",
    "pydantic==2.7.*",
    "qdrant-client==1.9.*",
    "anthropic==0.28.*",
    "sentence-transformers==3.0.*",
    "pypdf==4.2.*",
    "python-docx==1.1.*",
    "transformers==4.34.*",
    "httpx==0.27.*",
    "python-dotenv",
    "click"
]

[project.optional-dependencies]
test = [
    "pytest==8.0.*",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]